--- src/main/java/com/umc/hwaroak/config/security/jwt/JwtTokenProvider.java
@@ -1,69 +1,0 @@
-// JWT 발급, 검증, 파싱
-package com.umc.hwaroak.config.security.jwt;
-
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.security.Keys;
-
-import java.nio.charset.StandardCharsets;
-import java.security.Key;
-import java.util.Date;
-@Component
-public class JwtTokenProvider {
-
-    @Value("${jwt.secret}")
-    private String secretKey;
-
-    @Value("${jwt.access-token-validity}")
-    private long accessTokenValidity;
-
-    @Value("${jwt.refresh-token-validity}")
-    private long refreshTokenValidity;
-
-    // .yml에 설정한 jwt.secret을 바탕으로 서명키 생성(위조 방지)
-    private Key getSigningKey() {
-        return Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-    }
-
-    public String createAccessToken(Long userId) {
-        return createToken(userId, accessTokenValidity);
-    }
-
-    public String createRefreshToken(Long userId) {
-        return createToken(userId, refreshTokenValidity);
-    }
-
-    private String createToken(Long userId, long validity) {
-        Date now = new Date();
-        Key key = getSigningKey();
-
-        return Jwts.builder()
-                .setSubject(String.valueOf(userId))
-                .setIssuedAt(now)
-                .setExpiration(new Date(now.getTime() + validity))
-                .signWith(key, SignatureAlgorithm.HS256)
-                .compact();
-    }
-
-    // 토큰에서 사용자 id 추출
-    public Long getUserId(String token) {
-        Key key = getSigningKey();
-        Claims claims = Jwts.parser()
-                .setSigningKey(key)
-                .parseClaimsJws(token)
-                .getBody();
-        return Long.parseLong(claims.getSubject());
-    }
-
-    // 토큰 유효성 검증
-    public boolean validateToken(String token) {
-        Jwts.parser()
-                .setSigningKey(getSigningKey())
-                .parseClaimsJws(token);
-        return true;
-    }
-}
-
