+++ src/main/java/com/umc/hwaroak/exception/JwtExceptionHandler.java
@@ -5,13 +5,14 @@ import com.umc.hwaroak.response.ErrorResponse;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.MalformedJwtException;
 import io.jsonwebtoken.UnsupportedJwtException;
-import org.springframework.http.HttpStatus;
+import io.swagger.v3.oas.annotations.Hidden;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import java.security.SignatureException;
 @RestControllerAdvice
+@Hidden
 public class JwtExceptionHandler {
 
     @ExceptionHandler({SecurityException.class, SignatureException.class})
